Reset all shelves with button in SettingsView and SettingsViewTest

+++ src/main/java/com/karankumar/bookproject/backend/service/BookService.java
@@ -99,7 +99,14 @@ public class BookService extends BaseService<Book, Long> {
 
     @Override
     public void deleteAll() {
+        List<Book> books = bookRepository.findAll();
+        if(books.isEmpty()){
+            LOGGER.log(Level.INFO, "All books already deleted");
+            return;
+        }
+        LOGGER.log(Level.INFO, "Deleting all in books & authors. Book repository size = " + bookRepository.count());
         bookRepository.deleteAll();
         authorService.deleteAll();
+        LOGGER.log(Level.INFO, "Deleted all books in books & authors. Book repository size = " + bookRepository.count());
     }
 }
